pipeline/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
PyJWT[crypto]==2.12.1
